requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.2
django-debug-toolbar==3.7.0
exceptiongroup==1.0.1
factory-boy==3.2.1
Faker==15.3.1
iniconfig==1.1.1
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
psycopg2==2.9.5
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
toposort==1.7
